Bug 557313: cfx sdocs: create .md.json, don't archive ~ in docs tarball
This updates the "cfx sdocs" tool, which creates a tarball of docs files
(for serving from a static webserver, like apache). This tarball now
contains the generated/parsed JSON representation of the APIs, extracted
from the various markdown files. It also excludes emacs backup files,
which always have a tilde at the end of their filenames.