Modify on-demand copies of application files
The attempt is made to modify some of the linked files in the root
directory of a runnung application (i.e. build/name). However, the
original file a link is pointed to may not be write-accessible for
the running process (in the case the application path is passed from
one user to another, or it is a system-wide installation of an example
application). When that is the case, the symlink is substituted with
the write-accessible copy.